Basic AI finished
The simple AI is finished, you can find it in test_AI_v2 and v3. If you wish to experiment with it and push the changes, please use create a new v4 or something instead of altering the existing ones, as im still using those to test stuff (this includes AI_test.py).